Rename grid_search directory to _grid_search
